requirements.txt
openpyxl>=3.1.5
python-docx>=1.0.0
fpdf2>=2.8.0
